beta -> main 합병

+++ src/main/java/io/github/closeddev/Server/CreateServer.java
@@ -1,6 +1,5 @@
 package io.github.closeddev.Server;
 
-import io.github.closeddev.Class.SProperties;
 import io.github.closeddev.Downloader;
 import io.github.closeddev.JSONManager;
 import io.github.closeddev.Logger;
@@ -8,7 +7,6 @@ import io.github.closeddev.Main;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
-import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -17,22 +15,26 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 import static io.github.closeddev.Main.MCSBPath;
 
 public class CreateServer {
 
-    public static final String SERVERPATH = MCSBPath + "/Temp/testserver";
-    public static void createServer(String FullVersion, String Build, String vcode) {
+    public static HashMap<String, Object> PROPERTIES = new HashMap<>();
+
+    public static final String SERVERPATH = MCSBPath + "/Temp/Testserver";
+    public static void createServer(String FullVersion, String Build, String vcode, int ramAmount) {
         try {
             Main.makeDir(SERVERPATH);
 
             File verjson = new File(MCSBPath + "/versions.json");
             if (!verjson.isFile()) { // version.json 파일이 없을 경우 기본 파일 생성
                 JSONObject jsonObject = new JSONObject();
-                JSONObject asdf = new JSONObject(); // 이거슨 아무 의미도 없는 JSONObject 이름 원하시면 바꾸시오
-                asdf.put("list", new ArrayList<String>());
-                jsonObject.put("vers", asdf);
+                JSONObject versionObj = new JSONObject();
+                versionObj.put("list", new ArrayList<String>());
+                jsonObject.put("vers", versionObj);
                 JSONManager.writeJSON(MCSBPath + "/versions.json", jsonObject);
             }
 
@@ -68,29 +70,32 @@ public class CreateServer {
                 vers.put(FullVersion, versionfield);
             }
             JSONManager.writeJSON(MCSBPath + "/versions.json", jsonObject);
-            Files.copy(filename, Paths.get(SERVERPATH + "/paper.jar"));
+            if (!Paths.get(SERVERPATH + "/paper.jar").toFile().exists()) Files.copy(filename, Paths.get(SERVERPATH + "/paper.jar"));
             crteula(SERVERPATH);
-            crtstart(SERVERPATH);
-            SProperties properties = new SProperties();
+            crtstart(SERVERPATH, ramAmount);
+            crtproper();
+            Files.copy(Paths.get((MCSBPath + "/server.properties")), Paths.get(SERVERPATH + "/server.properties"));
+
 
 
         } catch(Exception e) {
             Logger.log(e.toString(), 1);
         }
     }
-    public static void crtstart(String svpath) {
+    public static void crtstart(String svpath, float ramAmount) {
         File BatFile = new File(svpath + "/start.bat");
         try {
             FileWriter fw = new FileWriter(BatFile, true);
-            fw.write("@echo off\njava -jar paper.jar\npause\nexit");
+            fw.write("@echo off\njava -" + "Xms" + ramAmount + "G -Xmx" + ramAmount + "G -jar paper.jar\npause\nexit");
             fw.flush();
             fw.close();
+            System.out.println("@echo off\njava -" + "Xms" + ramAmount + "G -Xmx" + ramAmount + "G -jar paper.jar\npause\nexit");
         } catch (IOException e) {
             Logger.log(e.toString(), 1);
         }
     }
 
-    public static void crteula(String svpath) {
+    private static void crteula(String svpath) {
         File BatFile = new File(svpath + "eula.txt");
         try {
             FileWriter fw = new FileWriter(BatFile, true);
@@ -100,5 +105,47 @@ public class CreateServer {
         } catch (IOException e) {
             Logger.log(e.toString(), 1);
         }
+    }
+
+    public static void crtproper() {
+
+        String configPath = MCSBPath + "/server.properties";
+        File config = new File(configPath);
+        if (!config.isFile()) {
+            try {
+                config.createNewFile();
+            } catch (IOException e) {
+                Logger.log("Failed to create server.properties File!", 1);
+            }
+        }
+
+        for (Map.Entry<String, Object> entry : PROPERTIES.entrySet()) {
+            String key = entry.getKey();
+            Object value = entry.getValue();
+
+            ServerManager.editProperties(configPath, key, value);
+        }
+    }
+
+    public static void setproperdefault() {
+        // Default properties value
+        PROPERTIES.put("level-seed", 0);
+        PROPERTIES.put("gamemode", "survival");
+        PROPERTIES.put("enable-command-block", false);
+        PROPERTIES.put("level-name", "world");
+        PROPERTIES.put("motd", "A Minecraft Server");
+        PROPERTIES.put("difficulty", "easy");
+        PROPERTIES.put("max-players", 20);
+        PROPERTIES.put("online-mode", true);
+        PROPERTIES.put("allow-flight", false);
+        PROPERTIES.put("allow-nether", true);
+        PROPERTIES.put("server-port", 25565);
+        PROPERTIES.put("hardcore", false);
+        PROPERTIES.put("white-list", false);
+        PROPERTIES.put("spawn-npcs", true);
+        PROPERTIES.put("spawn-animals", true);
+        PROPERTIES.put("level-type", "minecraft\\:default");
+        PROPERTIES.put("spawn-monsters", true);
+        PROPERTIES.put("spawn-protection", 16);
     }
 }
